+++ packages/producer/src/services/render/stages/captureStreamingStage.ts
@@ -58,11 +58,8 @@ import {
 } from "@hyperframes/engine";
 import type { FileServerHandle } from "../../fileServer.js";
 import type { ProducerLogger } from "../../../logger.js";
-import {
-  updateJobStatus,
-  type ProgressCallback,
-  type RenderJob,
-} from "../../renderOrchestrator.js";
+import type { ProgressCallback, RenderJob } from "../../renderOrchestrator.js";
+import { updateJobStatus } from "../shared.js";
 
 /**
  * Pre-built ffmpeg streaming-encoder options, exactly matching the
